moved common test framework
